Extract the `<BlockInfo />` container
