+++ packages/ui/src/common/components/BlockTime/BlockHistoryLine.tsx
@@ -1,30 +1,17 @@
 import React from 'react'
 import styled from 'styled-components'
 
-import { BlockIcon } from '@/common/components/icons'
 import { Tooltip, TooltipDefault } from '@/common/components/Tooltip'
 import { TextInlineSmall } from '@/common/components/typography'
 import { Fonts } from '@/common/constants'
-import { formatDateString, formatTokenValue } from '@/common/model/formatters'
-import { Block } from '@/common/types'
+import { formatDateString } from '@/common/model/formatters'
 
-import { BlockInfo, BlockNetworkInfo, BlockNumber } from './components'
+import { BlockInfo, BlockInfoProp } from './BlockInfo'
 
-export interface BlockHistoryLineProps {
-  block: Block
-}
-export const BlockHistoryLine = styled(({ block, ...props }: BlockHistoryLineProps) => (
+export const BlockHistoryLine = styled(({ block, ...props }: BlockInfoProp) => (
   <div {...props}>
     <TextInlineSmall lighter>{formatDateString(block.timestamp)}</TextInlineSmall>
-    <Tooltip
-      popupContent={
-        <BlockInfo>
-          <BlockIcon />
-          <BlockNumber>{formatTokenValue(block.number)}</BlockNumber>
-          <BlockNetworkInfo>on {block.network} network</BlockNetworkInfo>
-        </BlockInfo>
-      }
-    >
+    <Tooltip popupContent={<BlockInfo block={block} />}>
       <TooltipDefault />
     </Tooltip>
   </div>
